Rename TypeScript files and add config
