+++ Trader.ts
