+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "money-printer-bot",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "dist/server.js",
+  "scripts": {
+    "build": "tsc",
+    "start": "node dist/server.js"
+  },
+  "dependencies": {
+    "@solana/web3.js": "^1.88.2",
+    "axios": "^1.5.0",
+    "bs58": "^5.0.0",
+    "node-fetch": "^2.6.7",
+    "ws": "^8.15.1",
+    "express": "^4.18.2",
+    "cors": "^2.8.5",
+    "ioredis": "^5.3.2",
+    "@supabase/supabase-js": "^2.39.7",
+    "zod": "^3.22.4"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.0"
+  }
+}
